fix win build

+++ modules/guesser/src/guesser.cc
@@ -53,7 +53,7 @@ void guesser::init(motis::module::registry& reg) {
 
   auto stations = utl::to_vec(station_indices_, [&](unsigned i) {
     auto const& s = *sched.stations_[i];
-    double factor = 0;
+    float factor = 0;
     for (unsigned i = 0; i < s.dep_class_events_.size(); ++i) {
       factor += std::pow(10, (9 - i) / 3) * s.dep_class_events_.at(i);
     }
@@ -63,8 +63,8 @@ void guesser::init(motis::module::registry& reg) {
   if (!stations.empty()) {
     auto max_importatance =
         std::max_element(begin(stations), end(stations),
-                         [](std::pair<std::string, double> const& lhs,
-                            std::pair<std::string, double> const& rhs) {
+                         [](std::pair<std::string, float> const& lhs,
+                            std::pair<std::string, float> const& rhs) {
                            return lhs.second < rhs.second;
                          })
             ->second;
